loader: Fix failure on ICD loading
Piers (Nvidia) discovered a bug where the ICD would fail in different ways
depending on which ICD failed to load properly during the vkCreateInstance
call.  If the first ICD failed, but the second one succeeded, everything
would work fine, if the second failed, but the first succeeded, it would
fail completely because the last ICD failed.

Change-Id: I9f63146e60baf1ff310c948d8ef84c645519c075